feat: [IOCOM-2686] Remove UAL support for SEND dev env (#7552)
## Short description
This PR removes Universal (iOS) and App (Android) Link support for SEND
dev environment

## List of changes proposed in this pull request
- Removed dev SEND env from android manifest
- Removed dev SEND env from iOS entitlements

## How to test
On iOS, run the application on the simulator (or on a real device),
create a reminder (or a message) with the following linked url
`https://cittadini.dev.notifichedigitali.it/io?aar=asd123` and check
that, upon tapping the link, the IO App does **_NOT_** open.
In order to test this on Android, the build must be uploaded to the
stores, so that the auto-verification process can happen with the proper
signing certificate. As for this PR, check that the syntax is correct
with respect to the linked documentation and the associated online
files:
- [Asset Links -
Android](https://cittadini.dev.notifichedigitali.it/.well-known/assetlinks.json)
- [App Site Associations -
iOS](https://cittadini.dev.notifichedigitali.it/.well-known/apple-app-site-association)