chore: [IOBP-2136] add mixpanel idpay feature disable screen tracking (#7558)
## Short description
This pull request adds analytics tracking for when the entire IDPay
feature is disabled, ensuring that this user experience event is
properly logged.

## List of changes proposed in this pull request
* Added the `trackIDPayFeatureDisabledScreen` function to `index.ts` to
log the "IDPAY_FEATURE_DISABLED" event when the disabled screen is
shown.
* Imported and called `trackIDPayFeatureDisabledScreen` in
`IdPayDisabledScreen.tsx` so that the event is tracked on the first
render of the disabled screen.

## How to test
- On the dev-server, disable the entire `idPay` feature flag
- Check that whenever you try to start any IDPay-related flow, a
blocking screen appears and the mixpanel event is fired correctly.