fix(IT-Wallet): [SIW-3284] IT-Wallet credentials not accessible with poor network (#7564)
## Short description
This PR fixes a rare condition where the credential's details is not
accessible if the connection status is ok but we are unable to check if
the ITW remote flag is enabled

## List of changes proposed in this pull request
- Removed `ItwPresentationCredentialDetailScreen` from FF gate
- `ItwPresentationCredentialDetailScreen` hooks refactor 

## How to test
> [!WARNING]
> To test this PR you will need to mock the remote configuration
endpoint with a Proxy tool.

Verify that the credentials details screen is correctly accessible when
the app is online but the remote configuration is non available.
Also verify that no regressions are introduced when the app is offline.