fix: [IOPID-3529]: Polling in the email validation flow freezes when the app is in the background (#7561)
## Short description
This pull request enhances the `EmailValidationSendEmailScreen` by
improving how the app manages email validation polling based on the
app's foreground/background state. The main improvement is that polling
is now automatically paused when the app goes to the background and
resumed when the app returns to the foreground, which helps conserve
resources and provides a better user experience.

## List of changes proposed in this pull request
- Added an `AppState` event listener to pause polling when the app goes
to the background and resume polling when it returns to the foreground,
only if the email is not yet validated
(`EmailValidationSendEmailScreen.tsx`).
[[1]](diffhunk://#diff-be4f3b94c6a571c0669c20b47028fcca9bc73a88ffe35e8ceab0bd042e77f3cdL9-R9)
[[2]](diffhunk://#diff-be4f3b94c6a571c0669c20b47028fcca9bc73a88ffe35e8ceab0bd042e77f3cdR207-R222)

## Demo 
To see the demo video you can visit the [ticket page on
Jira](https://pagopa.atlassian.net/browse/IOPID-3529). The video
contains my sensitive data, so I prefer not to publish it here.


## How to test (ONLY ANDROID)
1. open app
2. change email
3. Receipt of validation email
4. Email validation
5. Return to app
6. Screen locks and won't proceed
7. App closes
8. App reopens
9. Normal access

Co-authored-by: Fabio Bombardi <16268789+shadowsheep1@users.noreply.github.com>