feat: [IOCOM-2717] Proper bottom sheet text when leaving an AAR notification (#7570)
## Short description
This PR fixes the text in the bottom sheet, when leaving an AAR send
notification, based on the user's mandate.

## List of changes proposed in this pull request
- Proper copy based on the mandate
- Mandate property from parent, to avoid duplicated logic

## How to test
Using the io-dev-api-server, check that the bottom sheet that is shown
when leaving an AAR notification has a different text, based on the fact
that the user has a permanent mandate or is the recipient.