revert: [IOBP-000] remove wallet update action from app startup (#7587)
## Short description
This pull request introduces a small change to the payments saga by
removing the wallet update action. The action was removed to reduce the
number of requests made at app startup.

However, this change introduces a regression affecting external deep
linking: when a deep link is triggered while the app is completely
closed, the data on the initial screens may fail to load properly,
resulting in an infinite loading state.

This behavior has been accepted for now, and a different solution will
be evaluated and implemented in the future to achieve a lighter and more
efficient approach.

## How to test
- Verify that when the app starts, wallet requests are not triggered
automatically, but are loaded only when the Wallet tab is opened.