chore: [IOAPPX-508] Remove javascript bundle reference on iOS (#7581)
## Short description
This PR removes the `main.jsbundle` file reference from the project. The
compiled JS bundle is no longer referenced in projects running on React
Native 0.78. Thus, removing it should be safe.

The other changes are generated automatically by Xcode when removing the
reference to `main.jsbundle`. The first one regarding the font seems to
be correct as other fonts included in this project have `9` as
`fileEncoding`. The other change regarding `TargetAttributes` seems to
be unrelated, however I'd leave as it is and see if it breaks something.
It's probably caused by the new Xcode version.

If the production builds fail we can revert this PR and then understand
why.

## List of changes proposed in this pull request
- Remove `main.jsbundle` reference from the project.

## How to test
This can be tested with the canary build number `3.19.0 (7663)` on
TestFlight and by trying to build the debug app on iOS.