chore: [IOBP-2180] add CdC wallet visibility remote feature flag selector (#7598)
## ⚠️ This PR depends on
https://github.com/pagopa/io-services-metadata/pull/1071

## Short description
This pull request introduces a new feature flag selector for controlling
the visibility of the CdC in the wallet section. It will be used in a
next iteration when implementing the entire CdC feature.

## List of changes proposed in this pull request
* Added `isCdCWalletVisibilityEnabledSelector` to `remoteConfig.ts` to
determine if the CdC wallet visibility feature is enabled based on
remote config and app version. This selector uses
`isPropertyWithMinAppVersionEnabled` and is intended for future use in
the implementation of the CdC wallet feature.
* Updated `IO_SERVICES_METADATA_VERSION` in
`scripts/generate-api-models.sh` to use the new version available of the
definitions.

## How to test
Static checks should resolve without any error