chore: [IOAPPX-510] Update import in the app review module on iOS (#7609)
## Short description
This PR updates the `SwiftUI` import in the `AppReviewModule` on iOS
which currently prevents building the app with Xcode 26.

## List of changes proposed in this pull request
- Replace `SwiftUICore` import with `SwiftUI`.

## How to test
Try to build the app on iOS with Xcode 26 or older versions if
available. From the debug menu you can test that the review still works
in the `App Feedback` section.