fix: [IOBP-2214] trigger the mixpanel event of the email notification banner only if banner shows (#7622)
## Short description
This pull request introduces a small change to the
`EmailNotificationBanner` component to improve its tracking logic. The
change ensures that the Mixpanel tracking event is only triggered when
the banner is eligible to be shown.

## List of changes proposed in this pull request
* The `useOnFirstRender` hook now takes a predicate function
(`canShowBanner`) as its second argument, so the Mixpanel tracking event
is only fired if the banner should be displayed.