feat(IT Wallet): [SIW-3274] Fetch digital credentials catalogue on app startup (#7607)
## Short description
This PR introduces the [digital credentials
catalogue](https://italia.github.io/eid-wallet-it-docs/releases/1.1.0/en/registry-catalogue.html).
When the app is launched the catalogue is fetched and persisted, so it
remains available in case of a failure the next time it is fetched.

## List of changes proposed in this pull request
- Added new reducer `credentialsCatalogue`
- Added a saga to fetch the catalogue on app startup
- Updated `@pagopa/io-react-native-wallet`

## How to test
Start the app, set the env to pre:
- Check the network and ensure there is a HTTP call to
`/.well-known/credential-catalogue`.
- Ensure `credentialsCatalogue` contains the catalogue.
- Mock the HTTP call to return a 500 error. The catalogue store should
still contain the previous persisted catalogue