feat(IT Wallet): [SIW-3038] Mixpanel Event for failed reissuing (#7578)
## Short description
This PR introduces tracking for the new event
`ITW_CREDENTIAL_REISSUING_FAILED` adding `itw_flow` and `reason`
properties.

## List of changes proposed in this pull request
- Added new properties: `ItwCredentialReissuingFailedProperties`
- Added a new object in `StoredCredential` type: `failure`
- Refactor of `setFailedCredential` in `itwCredentialUpgradeMachine`
- Added a new file `failure.ts` for the `itwCredentialUpgradeMachine` to
map the errors for the upgrade flow
- Added `trackItwCredentialReissuingFailed` in
`ItwIssuanceEidResultScreen`, the event is triggered if
`failedCredentials.length > 0 && failure`, for every failed Credential

## How to test
Perform an upgrade of credentials to L3 or a reissuing, verify on
Mixpanel that the event `ITW_CREDENTIAL_REISSUING_FAILED` is triggered
correctly if an error occurs.