refactor: [IOCOM-2801] Remove active session login guard and replace it with messages' saga guard (#7636)
## Short description
This PR refactors #7608, in order to move message's related business
logic into the same related section (instead of having it under the
login data).

## List of changes proposed in this pull request
- `refreshMessagesSection` has been removed from the authentication
section
- messages sagas are ready to run when property `messageSagasRegistered`
in reducer `sectionStatus` is `true`
- such property is set to `false` every time that the startup saga
restarts (by listening to `startApplicationInitialization`)
- such property is set to `true` when all messages' sagas have been
registered (responding to `setMessageSagasRegisteredAction` action,
which is dispatched at the end of the main messages' saga)
- `homeUtils` have been refactored to take into account such property's
value (instead of `refreshMessagesSection`)

## How to test
Using the io-dev-api-server, go to the Debug section and simulate an
active session login. Before pressing the login button, wait for 60
seconds to pass (which is the auto-refresh timer of the messages'
inbox). Perform the login and make sure that:
- messages' home has been refreshed properly
- manual refresh (of messages' home) works
- message's details can be properly loaded