feat(IT Wallet): [SIW-3241] Integration with the credentials catalogue for the field "Origine dei dati" (#7652)
## Short description
This PR removes the hardcoded `AuthSource` in favor of retrieving the
"Origine dei dati" field from the `CredentialsCatalogue`.

## List of changes proposed in this pull request
- Removed the hardcoded `authSource` from the translation JSON
- Added `itwCredentialByTypeSelector` to retrieve the correct credential
by searching for `credentialType`
- Added a utility function `mapCredentialsByType` to create a key-value
map for each credential
- Added `credentialFromCatalogue` in `ItwIssuanceMetadata`
- Modified `getItwAuthSource` to retrieve the `organization_name` field
- Removed `ItwCredentialAuthSource` as it is no longer necessary
- Fixed tests for `ItwPresentationClaimsSection` 

## How to test
Enable IT-Wallet, add a credential, navigate to the credential detail,
scroll to the bottom and verify that the "Origine dei dati" field is
correctly populated.