fix(IT Wallet): [SIW-3505] Make Android CIE read screen scrollable (#7706)
## Short description
This PR makes `ITW_IDENTIFICATION_CIE_AUTH_SCREEN` scrollable on
Android. On devices with smaller screens and the font size set to 200%
it was not possible to see all the content.

## List of changes proposed in this pull request
- Switched `View` with `ScrollView`

## How to test
Use an Android device with a small screen and set the system font to
200%. If not possible, edit the code to use a long placeholder text.

Ensure the CIE auth screen has not been modified with the default font
size.


https://github.com/user-attachments/assets/35a44c75-0ea1-4dd2-9609-d0784f9edb4e