feat(IT Wallet): [SIW-3460,SIW-3508] Improvement in case of status assertion errors (#7690)
## Short description
This PR introduces an improvement to the logic for retrieving the status
assertion: if AS or IPZS are down, we can still use the credentials up
to a maximum of 4 times. After that, the credentials will be shown as
not available.

> :warning: **UPDATE**: This PR introduces a workaround for SIW-3508,
allowing the app to request the `statusAssertion` on app restart
whenever the credential is expired or in an invalid/unknown state, in
order to recover from any temporary issue.

## List of changes proposed in this pull request
- Added `unverifiedCredentialsAccessCounter` redux state, with its
reducer
- Added `itwUnverifiedCredentialsCounterUp` and
`itwUnverifiedCredentialsCounterReset` actions
- Added `itwUnverifiedCredentialsCounterSelector` and
`itwUnverifiedCredentialsCounterLimitReached` selectors
- Added tests for the reducer
- Updated the `shouldRequestStatusAssertion` function: the credential is
now requested even when the status is invalid

## How to test
This feature can only be tested if IPZS or AS are down, or by forcing an
error in `verifyAndParseStatusAssertion` from the
`io-react-native-wallet` library.

## Preview

https://github.com/user-attachments/assets/34caca99-9b52-412e-be4a-7b63f1c838a8